+++ djangocms_alias/filters.py
@@ -31,3 +31,46 @@ class LanguageFilter(admin.SimpleListFilter):
                 ),
                 "display": title,
             }
+
+from django.contrib.sites.models import Site
+from cms.forms.utils import get_sites
+
+
+class SiteFilter(admin.SimpleListFilter):
+    title = _("Site")
+    parameter_name = "site"
+    no_site_set_value = "none"
+
+    def lookups(self, request, model_admin):
+        site_list = get_sites()
+        return [(site['id'], site['name']) for site in site_list]
+
+    def queryset(self, request, queryset):
+        chosen_site = self.value()
+        if chosen_site and chosen_site == self.no_site_set_value:
+            return queryset.filter(alias__site__isnull=True)
+        elif chosen_site:
+            return queryset.filter(alias__site=chosen_site)
+        return queryset
+
+    def choices(self, changelist):
+        yield {
+            "selected": self.value() is None,
+            "query_string": changelist.get_query_string(remove=[self.parameter_name]),
+            "display": _("Current"),
+        }
+        yield {
+            "selected": self.value() is self.no_site_set_value,
+            "query_string": changelist.get_query_string(
+                {self.parameter_name: self.no_site_set_value}
+            ),
+            "display": _("None"),
+        }
+        for lookup, title in self.lookup_choices:
+            yield {
+                "selected": self.value() == str(lookup),
+                "query_string": changelist.get_query_string(
+                    {self.parameter_name: lookup}
+                ),
+                "display": title,
+            }